Parallel video downloading implemented
So basically, with this new commit, the program can download multiple video at once. The reason I hadn't implemented this yes was because of an annoying bug the yt-dlp library contains. If you try to download multiple videos and you have hooked a progress hook on the, it tends to "spit" out irrelevant information about the video. A single if condition solved all this